autoedit: extract the code for tracking recent changes from recent edits retriever
